requirements.txt
# click==8.0.4
selenium==4.1.3
youtube-dl==2021.12.17
pyyaml==6.0